This is a great commit! Fixed a major bug where we would forget to flush out the blocked FragmentTaskMessages. This would be a problem for aborted transactions because the next transaction would execute it and would get incorrect results. Big win! I'm sitting here with Adriana. We just watched Ghostbusters. Now we're watching Louis in Miami!
git-svn-id: https://database.cs.brown.edu/svn/hstore/branches/newdtxn-branch@2530 7a8f3d5b-8c5f-44cb-9125-2414a256df87

+++ src/frontend/org/voltdb/messaging/FragmentTaskMessage.java
@@ -23,13 +23,11 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import org.apache.log4j.Logger;
 import org.voltdb.ParameterSet;
 import org.voltdb.VoltTable;
 import org.voltdb.utils.DBBPool;
-import org.voltdb.utils.Pair;
 
 import edu.mit.hstore.HStoreConstants;
 
@@ -250,8 +248,8 @@ public class FragmentTaskMessage extends TransactionInfoBaseMessage
             m_fragmentIds.length : 0;
     }
 
-    public long getFragmentId(int index) {
-        return m_fragmentIds[index];
+    public int getFragmentId(int index) {
+        return (int)m_fragmentIds[index];
     }
     
     public int getOutputDepId(int index) {
